session: migrate test-infra to testify for part of session_test
