+++ session/session_legacy_test.go
@@ -20,16 +20,13 @@ import (
 	"fmt"
 	"net"
 	"strconv"
-	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta/autoid"
@@ -46,7 +43,6 @@ import (
 	"github.com/pingcap/tidb/table/tables"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
-	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/testleak"
 	"github.com/tikv/client-go/v2/testutils"
@@ -2002,626 +1998,4 @@ func (s *testSessionSuite3) TestSetTransactionIsolationOneShot(c *C) {
 	tk.MustExec("begin")
 	_, err := tk.Se.Execute(ctx, "set transaction isolation level read committed")
 	c.Assert(err, NotNil)
-}
-
-func (s *testSessionSuite2) TestDBUserNameLength(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("create table if not exists t (a int)")
-	// Test user name length can be longer than 16.
-	tk.MustExec(`CREATE USER 'abcddfjakldfjaldddds'@'%' identified by ''`)
-	tk.MustExec(`grant all privileges on test.* to 'abcddfjakldfjaldddds'@'%'`)
-	tk.MustExec(`grant all privileges on test.t to 'abcddfjakldfjaldddds'@'%'`)
-}
-
-func (s *testSessionSuite2) TestHostLengthMax(c *C) {
-	host1 := strings.Repeat("a", 65)
-	host2 := strings.Repeat("a", 256)
-
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec(fmt.Sprintf(`CREATE USER 'abcddfjakldfjaldddds'@'%s'`, host1))
-
-	err := tk.ExecToErr(fmt.Sprintf(`CREATE USER 'abcddfjakldfjaldddds'@'%s'`, host2))
-	c.Assert(err.Error(), Equals, "[ddl:1470]String 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa' is too long for host name (should be no longer than 255)")
-}
-
-func (s *testSessionSuite2) TestCommitRetryCount(c *C) {
-	tk1 := testkit.NewTestKitWithInit(c, s.store)
-	tk2 := testkit.NewTestKitWithInit(c, s.store)
-	tk1.MustExec("create table no_retry (id int)")
-	tk1.MustExec("insert into no_retry values (1)")
-	tk1.MustExec("set @@tidb_retry_limit = 0")
-
-	tk1.MustExec("begin")
-	tk1.MustExec("update no_retry set id = 2")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update no_retry set id = 3")
-	tk2.MustExec("commit")
-
-	// No auto retry because retry limit is set to 0.
-	_, err := tk1.Se.Execute(context.Background(), "commit")
-	c.Assert(err, NotNil)
-}
-
-func (s *testSessionSuite3) TestEnablePartition(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("set tidb_enable_table_partition=off")
-	tk.MustQuery("show variables like 'tidb_enable_table_partition'").Check(testkit.Rows("tidb_enable_table_partition OFF"))
-
-	tk.MustExec("set global tidb_enable_table_partition = on")
-
-	tk.MustQuery("show variables like 'tidb_enable_table_partition'").Check(testkit.Rows("tidb_enable_table_partition OFF"))
-	tk.MustQuery("show global variables like 'tidb_enable_table_partition'").Check(testkit.Rows("tidb_enable_table_partition ON"))
-
-	tk.MustExec("set tidb_enable_list_partition=off")
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition OFF"))
-	tk.MustExec("set global tidb_enable_list_partition=on")
-	tk.MustQuery("show global variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition OFF"))
-
-	tk.MustExec("set tidb_enable_list_partition=1")
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-
-	tk.MustExec("set tidb_enable_list_partition=on")
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-
-	tk.MustQuery("show global variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-	tk.MustExec("set global tidb_enable_list_partition=off")
-	tk.MustQuery("show global variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition OFF"))
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-	tk.MustExec("set tidb_enable_list_partition=off")
-	tk.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition OFF"))
-
-	tk.MustExec("set global tidb_enable_list_partition=on")
-	tk.MustQuery("show global variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-	tk1 := testkit.NewTestKitWithInit(c, s.store)
-	tk1.MustQuery("show variables like 'tidb_enable_table_partition'").Check(testkit.Rows("tidb_enable_table_partition ON"))
-	tk1.MustQuery("show variables like 'tidb_enable_list_partition'").Check(testkit.Rows("tidb_enable_list_partition ON"))
-}
-
-func (s *testSessionSuite2) TestUpdatePrivilege(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists t1, t2;")
-	tk.MustExec("create table t1 (id int);")
-	tk.MustExec("create table t2 (id int);")
-	tk.MustExec("insert into t1 values (1);")
-	tk.MustExec("insert into t2 values (2);")
-	tk.MustExec("create user xxx;")
-	tk.MustExec("grant all on test.t1 to xxx;")
-	tk.MustExec("grant select on test.t2 to xxx;")
-
-	tk1 := testkit.NewTestKitWithInit(c, s.store)
-	c.Assert(tk1.Se.Auth(&auth.UserIdentity{Username: "xxx", Hostname: "localhost"},
-		[]byte(""),
-		[]byte("")), IsTrue)
-
-	_, err := tk1.Exec("update t2 set id = 666 where id = 1;")
-	c.Assert(err, NotNil)
-	c.Assert(strings.Contains(err.Error(), "privilege check"), IsTrue)
-
-	// Cover a bug that t1 and t2 both require update privilege.
-	// In fact, the privlege check for t1 should be update, and for t2 should be select.
-	_, err = tk1.Exec("update t1,t2 set t1.id = t2.id;")
-	c.Assert(err, IsNil)
-
-	// Fix issue 8911
-	tk.MustExec("create database weperk")
-	tk.MustExec("use weperk")
-	tk.MustExec("create table tb_wehub_server (id int, active_count int, used_count int)")
-	tk.MustExec("create user 'weperk'")
-	tk.MustExec("grant all privileges on weperk.* to 'weperk'@'%'")
-	c.Assert(tk1.Se.Auth(&auth.UserIdentity{Username: "weperk", Hostname: "%"},
-		[]byte(""), []byte("")), IsTrue)
-	tk1.MustExec("use weperk")
-	tk1.MustExec("update tb_wehub_server a set a.active_count=a.active_count+1,a.used_count=a.used_count+1 where id=1")
-
-	tk.MustExec("create database service")
-	tk.MustExec("create database report")
-	tk.MustExec(`CREATE TABLE service.t1 (
-  id int(11) DEFAULT NULL,
-  a bigint(20) NOT NULL,
-  b text DEFAULT NULL,
-  PRIMARY KEY (a)
-)`)
-	tk.MustExec(`CREATE TABLE report.t2 (
-  a bigint(20) DEFAULT NULL,
-  c bigint(20) NOT NULL
-)`)
-	tk.MustExec("grant all privileges on service.* to weperk")
-	tk.MustExec("grant all privileges on report.* to weperk")
-	tk1.Se.GetSessionVars().CurrentDB = ""
-	tk1.MustExec(`update service.t1 s,
-report.t2 t
-set s.a = t.a
-WHERE
-s.a = t.a
-and t.c >=  1 and t.c <= 10000
-and s.b !='xx';`)
-
-	// Fix issue 10028
-	tk.MustExec("create database ap")
-	tk.MustExec("create database tp")
-	tk.MustExec("grant all privileges on ap.* to xxx")
-	tk.MustExec("grant select on tp.* to xxx")
-	tk.MustExec("create table tp.record( id int,name varchar(128),age int)")
-	tk.MustExec("insert into tp.record (id,name,age) values (1,'john',18),(2,'lary',19),(3,'lily',18)")
-	tk.MustExec("create table ap.record( id int,name varchar(128),age int)")
-	tk.MustExec("insert into ap.record(id) values(1)")
-	c.Assert(tk1.Se.Auth(&auth.UserIdentity{Username: "xxx", Hostname: "localhost"},
-		[]byte(""),
-		[]byte("")), IsTrue)
-	_, err2 := tk1.Exec("update ap.record t inner join tp.record tt on t.id=tt.id  set t.name=tt.name")
-	c.Assert(err2, IsNil)
-}
-
-func (s *testSessionSuite2) TestTxnGoString(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists gostr;")
-	tk.MustExec("create table gostr (id int);")
-	txn, err := tk.Se.Txn(false)
-	c.Assert(err, IsNil)
-	str1 := fmt.Sprintf("%#v", txn)
-	c.Assert(str1, Equals, "Txn{state=invalid}")
-	tk.MustExec("begin")
-	txn, err = tk.Se.Txn(false)
-	c.Assert(err, IsNil)
-	c.Assert(fmt.Sprintf("%#v", txn), Equals, fmt.Sprintf("Txn{state=valid, txnStartTS=%d}", txn.StartTS()))
-
-	tk.MustExec("insert into gostr values (1)")
-	c.Assert(fmt.Sprintf("%#v", txn), Equals, fmt.Sprintf("Txn{state=valid, txnStartTS=%d}", txn.StartTS()))
-
-	tk.MustExec("rollback")
-	c.Assert(fmt.Sprintf("%#v", txn), Equals, "Txn{state=invalid}")
-}
-
-func (s *testSessionSuite3) TestMaxExeucteTime(c *C) {
-	var err error
-	tk := testkit.NewTestKit(c, s.store)
-	tk.Se, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-
-	tk.MustExec("use test")
-	tk.MustExec("create table MaxExecTime( id int,name varchar(128),age int);")
-	tk.MustExec("begin")
-	tk.MustExec("insert into MaxExecTime (id,name,age) values (1,'john',18),(2,'lary',19),(3,'lily',18);")
-
-	tk.MustQuery("select /*+ MAX_EXECUTION_TIME(1000) MAX_EXECUTION_TIME(500) */ * FROM MaxExecTime;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings()[0].Err.Error(), Equals, "MAX_EXECUTION_TIME() is defined more than once, only the last definition takes effect: MAX_EXECUTION_TIME(500)")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.HasMaxExecutionTime, Equals, true)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MaxExecutionTime, Equals, uint64(500))
-
-	tk.MustQuery("select @@MAX_EXECUTION_TIME;").Check(testkit.Rows("0"))
-	tk.MustQuery("select @@global.MAX_EXECUTION_TIME;").Check(testkit.Rows("0"))
-	tk.MustQuery("select /*+ MAX_EXECUTION_TIME(1000) */ * FROM MaxExecTime;")
-
-	tk.MustExec("set @@global.MAX_EXECUTION_TIME = 300;")
-	tk.MustQuery("select * FROM MaxExecTime;")
-
-	tk.MustExec("set @@MAX_EXECUTION_TIME = 150;")
-	tk.MustQuery("select * FROM MaxExecTime;")
-
-	tk.MustQuery("select @@global.MAX_EXECUTION_TIME;").Check(testkit.Rows("300"))
-	tk.MustQuery("select @@MAX_EXECUTION_TIME;").Check(testkit.Rows("150"))
-
-	tk.MustExec("set @@global.MAX_EXECUTION_TIME = 0;")
-	tk.MustExec("set @@MAX_EXECUTION_TIME = 0;")
-	tk.MustExec("commit")
-	tk.MustExec("drop table if exists MaxExecTime;")
-}
-
-func (s *testSessionSuite2) TestGrantViewRelated(c *C) {
-	tkRoot := testkit.NewTestKitWithInit(c, s.store)
-	tkUser := testkit.NewTestKitWithInit(c, s.store)
-
-	tkRoot.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "localhost", CurrentUser: true, AuthUsername: "root", AuthHostname: "%"}, nil, []byte("012345678901234567890"))
-
-	tkRoot.MustExec("create table if not exists t (a int)")
-	tkRoot.MustExec("create view v_version29 as select * from t")
-	tkRoot.MustExec("create user 'u_version29'@'%'")
-	tkRoot.MustExec("grant select on t to u_version29@'%'")
-
-	tkUser.Se.Auth(&auth.UserIdentity{Username: "u_version29", Hostname: "localhost", CurrentUser: true, AuthUsername: "u_version29", AuthHostname: "%"}, nil, []byte("012345678901234567890"))
-
-	tkUser.MustQuery("select current_user();").Check(testkit.Rows("u_version29@%"))
-	err := tkUser.ExecToErr("select * from test.v_version29;")
-	c.Assert(err, NotNil)
-	tkUser.MustQuery("select current_user();").Check(testkit.Rows("u_version29@%"))
-	err = tkUser.ExecToErr("create view v_version29_c as select * from t;")
-	c.Assert(err, NotNil)
-
-	tkRoot.MustExec(`grant show view, select on v_version29 to 'u_version29'@'%'`)
-	tkRoot.MustQuery("select table_priv from mysql.tables_priv where host='%' and db='test' and user='u_version29' and table_name='v_version29'").Check(testkit.Rows("Select,Show View"))
-
-	tkUser.MustQuery("select current_user();").Check(testkit.Rows("u_version29@%"))
-	tkUser.MustQuery("show create view v_version29;")
-	err = tkUser.ExecToErr("create view v_version29_c as select * from v_version29;")
-	c.Assert(err, NotNil)
-
-	tkRoot.MustExec("create view v_version29_c as select * from v_version29;")
-	tkRoot.MustExec(`grant create view on v_version29_c to 'u_version29'@'%'`) // Can't grant privilege on a non-exist table/view.
-	tkRoot.MustQuery("select table_priv from mysql.tables_priv where host='%' and db='test' and user='u_version29' and table_name='v_version29_c'").Check(testkit.Rows("Create View"))
-	tkRoot.MustExec("drop view v_version29_c")
-
-	tkRoot.MustExec(`grant select on v_version29 to 'u_version29'@'%'`)
-	tkUser.MustQuery("select current_user();").Check(testkit.Rows("u_version29@%"))
-	tkUser.MustExec("create view v_version29_c as select * from v_version29;")
-}
-
-func (s *testSessionSuite3) TestLoadClientInteractive(c *C) {
-	var (
-		err          error
-		connectionID uint64
-	)
-	tk := testkit.NewTestKit(c, s.store)
-	tk.Se, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-	id := atomic.AddUint64(&connectionID, 1)
-	tk.Se.SetConnectionID(id)
-	tk.Se.GetSessionVars().ClientCapability = tk.Se.GetSessionVars().ClientCapability | mysql.ClientInteractive
-	tk.MustQuery("select @@wait_timeout").Check(testkit.Rows("28800"))
-}
-
-func (s *testSessionSuite2) TestReplicaRead(c *C) {
-	var err error
-	tk := testkit.NewTestKit(c, s.store)
-	tk.Se, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-	c.Assert(tk.Se.GetSessionVars().GetReplicaRead(), Equals, kv.ReplicaReadLeader)
-	tk.MustExec("set @@tidb_replica_read = 'follower';")
-	c.Assert(tk.Se.GetSessionVars().GetReplicaRead(), Equals, kv.ReplicaReadFollower)
-	tk.MustExec("set @@tidb_replica_read = 'leader';")
-	c.Assert(tk.Se.GetSessionVars().GetReplicaRead(), Equals, kv.ReplicaReadLeader)
-}
-
-func (s *testSessionSuite3) TestIsolationRead(c *C) {
-	var err error
-	tk := testkit.NewTestKit(c, s.store)
-	tk.Se, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-	c.Assert(len(tk.Se.GetSessionVars().GetIsolationReadEngines()), Equals, 3)
-	tk.MustExec("set @@tidb_isolation_read_engines = 'tiflash';")
-	engines := tk.Se.GetSessionVars().GetIsolationReadEngines()
-	c.Assert(len(engines), Equals, 1)
-	_, hasTiFlash := engines[kv.TiFlash]
-	_, hasTiKV := engines[kv.TiKV]
-	c.Assert(hasTiFlash, Equals, true)
-	c.Assert(hasTiKV, Equals, false)
-}
-
-func (s *testSessionSuite2) TestStmtHints(c *C) {
-	var err error
-	tk := testkit.NewTestKit(c, s.store)
-	tk.Se, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-
-	// Test MEMORY_QUOTA hint
-	tk.MustExec("select /*+ MEMORY_QUOTA(1 MB) */ 1;")
-	val := int64(1) * 1024 * 1024
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-	tk.MustExec("select /*+ MEMORY_QUOTA(1 GB) */ 1;")
-	val = int64(1) * 1024 * 1024 * 1024
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-	tk.MustExec("select /*+ MEMORY_QUOTA(1 GB), MEMORY_QUOTA(1 MB) */ 1;")
-	val = int64(1) * 1024 * 1024
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-	tk.MustExec("select /*+ MEMORY_QUOTA(0 GB) */ 1;")
-	val = int64(0)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings()[0].Err.Error(), Equals, "Setting the MEMORY_QUOTA to 0 means no memory limit")
-
-	tk.MustExec("use test")
-	tk.MustExec("create table t1(a int);")
-	tk.MustExec("insert /*+ MEMORY_QUOTA(1 MB) */ into t1 (a) values (1);")
-	val = int64(1) * 1024 * 1024
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-
-	tk.MustExec("insert /*+ MEMORY_QUOTA(1 MB) */  into t1 select /*+ MEMORY_QUOTA(3 MB) */ * from t1;")
-	val = int64(1) * 1024 * 1024
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.MemTracker.CheckBytesLimit(val), IsTrue)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings()[0].Err.Error(), Equals, "[util:3126]Hint MEMORY_QUOTA(`3145728`) is ignored as conflicting/duplicated.")
-
-	// Test NO_INDEX_MERGE hint
-	tk.Se.GetSessionVars().SetEnableIndexMerge(true)
-	tk.MustExec("select /*+ NO_INDEX_MERGE() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.NoIndexMergeHint, IsTrue)
-	tk.MustExec("select /*+ NO_INDEX_MERGE(), NO_INDEX_MERGE() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().GetEnableIndexMerge(), IsTrue)
-
-	// Test STRAIGHT_JOIN hint
-	tk.MustExec("select /*+ straight_join() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.StraightJoinOrder, IsTrue)
-	tk.MustExec("select /*+ straight_join(), straight_join() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-
-	// Test USE_TOJA hint
-	tk.Se.GetSessionVars().SetAllowInSubqToJoinAndAgg(true)
-	tk.MustExec("select /*+ USE_TOJA(false) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().GetAllowInSubqToJoinAndAgg(), IsFalse)
-	tk.Se.GetSessionVars().SetAllowInSubqToJoinAndAgg(false)
-	tk.MustExec("select /*+ USE_TOJA(true) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().GetAllowInSubqToJoinAndAgg(), IsTrue)
-	tk.MustExec("select /*+ USE_TOJA(false), USE_TOJA(true) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().GetAllowInSubqToJoinAndAgg(), IsTrue)
-
-	// Test USE_CASCADES hint
-	tk.Se.GetSessionVars().SetEnableCascadesPlanner(true)
-	tk.MustExec("select /*+ USE_CASCADES(false) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().GetEnableCascadesPlanner(), IsFalse)
-	tk.Se.GetSessionVars().SetEnableCascadesPlanner(false)
-	tk.MustExec("select /*+ USE_CASCADES(true) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().GetEnableCascadesPlanner(), IsTrue)
-	tk.MustExec("select /*+ USE_CASCADES(false), USE_CASCADES(true) */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings()[0].Err.Error(), Equals, "USE_CASCADES() is defined more than once, only the last definition takes effect: USE_CASCADES(true)")
-	c.Assert(tk.Se.GetSessionVars().GetEnableCascadesPlanner(), IsTrue)
-
-	// Test READ_CONSISTENT_REPLICA hint
-	tk.Se.GetSessionVars().SetReplicaRead(kv.ReplicaReadLeader)
-	tk.MustExec("select /*+ READ_CONSISTENT_REPLICA() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().GetReplicaRead(), Equals, kv.ReplicaReadFollower)
-	tk.MustExec("select /*+ READ_CONSISTENT_REPLICA(), READ_CONSISTENT_REPLICA() */ 1;")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 1)
-	c.Assert(tk.Se.GetSessionVars().GetReplicaRead(), Equals, kv.ReplicaReadFollower)
-}
-
-func (s *testSessionSuite3) TestPessimisticLockOnPartition(c *C) {
-	// This test checks that 'select ... for update' locks the partition instead of the table.
-	// Cover a bug that table ID is used to encode the lock key mistakenly.
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec(`create table if not exists forupdate_on_partition (
-  age int not null primary key,
-  nickname varchar(20) not null,
-  gender int not null default 0,
-  first_name varchar(30) not null default '',
-  last_name varchar(20) not null default '',
-  full_name varchar(60) as (concat(first_name, ' ', last_name)),
-  index idx_nickname (nickname)
-) partition by range (age) (
-  partition child values less than (18),
-  partition young values less than (30),
-  partition middle values less than (50),
-  partition old values less than (123)
-);`)
-	tk.MustExec("insert into forupdate_on_partition (`age`, `nickname`) values (25, 'cosven');")
-
-	tk1 := testkit.NewTestKit(c, s.store)
-	tk1.MustExec("use test")
-
-	tk.MustExec("begin pessimistic")
-	tk.MustQuery("select * from forupdate_on_partition where age=25 for update").Check(testkit.Rows("25 cosven 0    "))
-	tk1.MustExec("begin pessimistic")
-
-	ch := make(chan int32, 5)
-	go func() {
-		tk1.MustExec("update forupdate_on_partition set first_name='sw' where age=25")
-		ch <- 0
-		tk1.MustExec("commit")
-		ch <- 0
-	}()
-
-	// Leave 50ms for tk1 to run, tk1 should be blocked at the update operation.
-	time.Sleep(50 * time.Millisecond)
-	ch <- 1
-
-	tk.MustExec("commit")
-	// tk1 should be blocked until tk commit, check the order.
-	c.Assert(<-ch, Equals, int32(1))
-	c.Assert(<-ch, Equals, int32(0))
-	<-ch // wait for goroutine to quit.
-
-	// Once again...
-	// This time, test for the update-update conflict.
-	tk.MustExec("begin pessimistic")
-	tk.MustExec("update forupdate_on_partition set first_name='sw' where age=25")
-	tk1.MustExec("begin pessimistic")
-
-	go func() {
-		tk1.MustExec("update forupdate_on_partition set first_name = 'xxx' where age=25")
-		ch <- 0
-		tk1.MustExec("commit")
-		ch <- 0
-	}()
-
-	// Leave 50ms for tk1 to run, tk1 should be blocked at the update operation.
-	time.Sleep(50 * time.Millisecond)
-	ch <- 1
-
-	tk.MustExec("commit")
-	// tk1 should be blocked until tk commit, check the order.
-	c.Assert(<-ch, Equals, int32(1))
-	c.Assert(<-ch, Equals, int32(0))
-	<-ch // wait for goroutine to quit.
-}
-
-func (s *testSessionSuite2) TestPerStmtTaskID(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("create table task_id (v int)")
-
-	tk.MustExec("begin")
-	tk.MustExec("select * from task_id where v > 10")
-	taskID1 := tk.Se.GetSessionVars().StmtCtx.TaskID
-	tk.MustExec("select * from task_id where v < 5")
-	taskID2 := tk.Se.GetSessionVars().StmtCtx.TaskID
-	tk.MustExec("commit")
-
-	c.Assert(taskID1 != taskID2, IsTrue)
-}
-
-func (s *testSessionSuite2) TestSetEnableRateLimitAction(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	// assert default value
-	result := tk.MustQuery("select @@tidb_enable_rate_limit_action;")
-	result.Check(testkit.Rows("1"))
-	tk.MustExec("use test")
-	tk.MustExec("create table tmp123(id int)")
-	tk.MustQuery("select * from tmp123;")
-	haveRateLimitAction := false
-	action := tk.Se.GetSessionVars().StmtCtx.MemTracker.GetFallbackForTest(false)
-	for ; action != nil; action = action.GetFallback() {
-		if action.GetPriority() == memory.DefRateLimitPriority {
-			haveRateLimitAction = true
-			break
-		}
-	}
-	c.Assert(haveRateLimitAction, IsTrue)
-
-	// assert set sys variable
-	tk.MustExec("set global tidb_enable_rate_limit_action= '0';")
-	tk.Se.Close()
-
-	se, err := session.CreateSession4Test(s.store)
-	c.Check(err, IsNil)
-	tk.Se = se
-	result = tk.MustQuery("select @@tidb_enable_rate_limit_action;")
-	result.Check(testkit.Rows("0"))
-
-	haveRateLimitAction = false
-	action = tk.Se.GetSessionVars().StmtCtx.MemTracker.GetFallbackForTest(false)
-	for ; action != nil; action = action.GetFallback() {
-		if action.GetPriority() == memory.DefRateLimitPriority {
-			haveRateLimitAction = true
-			break
-		}
-	}
-	c.Assert(haveRateLimitAction, IsFalse)
-}
-
-func (s *testSessionSuite2) TestIssue19127(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-	tk.MustExec("drop table if exists issue19127")
-	tk.MustExec("create table issue19127 (c_int int, c_str varchar(40), primary key (c_int, c_str) ) partition by hash (c_int) partitions 4;")
-	tk.MustExec("insert into issue19127 values (9, 'angry williams'), (10, 'thirsty hugle');")
-	tk.Exec("update issue19127 set c_int = c_int + 10, c_str = 'adoring stonebraker' where c_int in (10, 9);")
-	c.Assert(tk.Se.AffectedRows(), Equals, uint64(2))
-}
-
-func (s *testSessionSuite2) TestMemoryUsageAlarmVariable(c *C) {
-	tk := testkit.NewTestKitWithInit(c, s.store)
-
-	tk.MustExec("set @@global.tidb_memory_usage_alarm_ratio=1")
-	tk.MustQuery("select @@global.tidb_memory_usage_alarm_ratio").Check(testkit.Rows("1"))
-	tk.MustExec("set @@global.tidb_memory_usage_alarm_ratio=0")
-	tk.MustQuery("select @@global.tidb_memory_usage_alarm_ratio").Check(testkit.Rows("0"))
-	tk.MustExec("set @@global.tidb_memory_usage_alarm_ratio=0.7")
-	tk.MustQuery("select @@global.tidb_memory_usage_alarm_ratio").Check(testkit.Rows("0.7"))
-	tk.MustExec("set @@global.tidb_memory_usage_alarm_ratio=1.1")
-	tk.MustQuery("SHOW WARNINGS").Check(testkit.Rows("Warning 1292 Truncated incorrect tidb_memory_usage_alarm_ratio value: '1.1'"))
-	tk.MustQuery("select @@global.tidb_memory_usage_alarm_ratio").Check(testkit.Rows("1"))
-
-	tk.MustExec("set @@global.tidb_memory_usage_alarm_ratio=-1")
-	tk.MustQuery("SHOW WARNINGS").Check(testkit.Rows("Warning 1292 Truncated incorrect tidb_memory_usage_alarm_ratio value: '-1'"))
-	tk.MustQuery("select @@global.tidb_memory_usage_alarm_ratio").Check(testkit.Rows("0"))
-
-	tk.MustExec("set @@session.tidb_memory_usage_alarm_ratio=0.8")
-	tk.MustQuery(`show warnings`).Check(testkit.Rows(fmt.Sprintf("Warning %d modifying tidb_memory_usage_alarm_ratio will require SET GLOBAL in a future version of TiDB", errno.ErrInstanceScope)))
-}
-
-func (s *testSessionSuite2) TestSelectLockInShare(c *C) {
-	tk1 := testkit.NewTestKitWithInit(c, s.store)
-	tk1.MustExec("DROP TABLE IF EXISTS t_sel_in_share")
-	tk1.MustExec("CREATE TABLE t_sel_in_share (id int DEFAULT NULL)")
-	tk1.MustExec("insert into t_sel_in_share values (11)")
-	err := tk1.ExecToErr("select * from t_sel_in_share lock in share mode")
-	c.Assert(err, NotNil)
-	tk1.MustExec("set @@tidb_enable_noop_functions = 1")
-	tk1.MustQuery("select * from t_sel_in_share lock in share mode").Check(testkit.Rows("11"))
-	tk1.MustExec("DROP TABLE t_sel_in_share")
-}
-
-func (s *testSessionSuite) TestReadDMLBatchSize(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("set global tidb_dml_batch_size=1000")
-	se, err := session.CreateSession(s.store)
-	c.Assert(err, IsNil)
-	// `select 1` to load the global variables.
-	_, _ = se.Execute(context.TODO(), "select 1")
-	c.Assert(se.GetSessionVars().DMLBatchSize, Equals, 1000)
-}
-
-func (s *testSessionSuite) TestInTxnPSProtoPointGet(c *C) {
-	ctx := context.Background()
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("create table t1(c1 int primary key, c2 int, c3 int)")
-	tk.MustExec("insert into t1 values(1, 10, 100)")
-
-	// Generate the ps statement and make the prepared plan cached for point get.
-	id, _, _, err := tk.Se.PrepareStmt("select c1, c2 from t1 where c1 = ?")
-	c.Assert(err, IsNil)
-	idForUpdate, _, _, err := tk.Se.PrepareStmt("select c1, c2 from t1 where c1 = ? for update")
-	c.Assert(err, IsNil)
-	params := []types.Datum{types.NewDatum(1)}
-	rs, err := tk.Se.ExecutePreparedStmt(ctx, id, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, idForUpdate, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-
-	// Query again the cached plan will be used.
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, id, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, idForUpdate, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-
-	// Start a transaction, now the in txn flag will be added to the session vars.
-	_, err = tk.Se.Execute(ctx, "start transaction")
-	c.Assert(err, IsNil)
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, id, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-	txn, err := tk.Se.Txn(false)
-	c.Assert(err, IsNil)
-	c.Assert(txn.Valid(), IsTrue)
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, idForUpdate, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 10"))
-	txn, err = tk.Se.Txn(false)
-	c.Assert(err, IsNil)
-	c.Assert(txn.Valid(), IsTrue)
-	_, err = tk.Se.Execute(ctx, "update t1 set c2 = c2 + 1")
-	c.Assert(err, IsNil)
-	// Check the read result after in-transaction update.
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, id, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 11"))
-	rs, err = tk.Se.ExecutePreparedStmt(ctx, idForUpdate, params)
-	c.Assert(err, IsNil)
-	tk.ResultSetToResult(rs, Commentf("%v", rs)).Check(testkit.Rows("1 11"))
-	txn, err = tk.Se.Txn(false)
-	c.Assert(err, IsNil)
-	c.Assert(txn.Valid(), IsTrue)
-	tk.MustExec("commit")
-}
-
-func (s *testSessionSuite) TestAuthPluginForUser(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("CREATE USER 'tapfu1' IDENTIFIED WITH mysql_native_password BY 'tapfu1'")
-	plugin, err := tk.Se.AuthPluginForUser(&auth.UserIdentity{Username: "tapfu1", Hostname: `%`})
-	c.Assert(err, IsNil)
-	c.Assert(plugin, Equals, "mysql_native_password")
-
-	tk.MustExec("CREATE USER 'tapfu2' IDENTIFIED WITH mysql_native_password")
-	plugin, err = tk.Se.AuthPluginForUser(&auth.UserIdentity{Username: "tapfu2", Hostname: `%`})
-	c.Assert(err, IsNil)
-	c.Assert(plugin, Equals, "")
-
-	tk.MustExec("CREATE USER 'tapfu3' IDENTIFIED WITH caching_sha2_password BY 'tapfu3'")
-	plugin, err = tk.Se.AuthPluginForUser(&auth.UserIdentity{Username: "tapfu3", Hostname: `%`})
-	c.Assert(err, IsNil)
-	c.Assert(plugin, Equals, "caching_sha2_password")
-
-	tk.MustExec("CREATE USER 'tapfu4' IDENTIFIED WITH caching_sha2_password")
-	plugin, err = tk.Se.AuthPluginForUser(&auth.UserIdentity{Username: "tapfu4", Hostname: `%`})
-	c.Assert(err, IsNil)
-	c.Assert(plugin, Equals, "")
 }
