+++ tests/realtikvtest/txntest/txn_test.go
@@ -1,0 +1,114 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package txntest
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/pingcap/tidb/testkit"
+	"github.com/pingcap/tidb/tests/realtikvtest"
+	"github.com/pingcap/tidb/types"
+	"github.com/stretchr/testify/require"
+)
+
+func TestInTxnPSProtoPointGet(t *testing.T) {
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("create table t1(c1 int primary key, c2 int, c3 int)")
+	tk.MustExec("insert into t1 values(1, 10, 100)")
+
+	ctx := context.Background()
+
+	// Generate the ps statement and make the prepared plan cached for point get.
+	id, _, _, err := tk.Session().PrepareStmt("select c1, c2 from t1 where c1 = ?")
+	require.NoError(t, err)
+	idForUpdate, _, _, err := tk.Session().PrepareStmt("select c1, c2 from t1 where c1 = ? for update")
+	require.NoError(t, err)
+	params := []types.Datum{types.NewDatum(1)}
+	rs, err := tk.Session().ExecutePreparedStmt(ctx, id, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, idForUpdate, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+
+	// Query again the cached plan will be used.
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, id, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, idForUpdate, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+
+	// Start a transaction, now the in txn flag will be added to the session vars.
+	_, err = tk.Session().Execute(ctx, "start transaction")
+	require.NoError(t, err)
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, id, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+	txn, err := tk.Session().Txn(false)
+	require.NoError(t, err)
+	require.True(t, txn.Valid())
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, idForUpdate, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 10"))
+	txn, err = tk.Session().Txn(false)
+	require.NoError(t, err)
+	require.True(t, txn.Valid())
+	_, err = tk.Session().Execute(ctx, "update t1 set c2 = c2 + 1")
+	require.NoError(t, err)
+	// Check the read result after in-transaction update.
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, id, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 11"))
+	rs, err = tk.Session().ExecutePreparedStmt(ctx, idForUpdate, params)
+	require.NoError(t, err)
+	tk.ResultSetToResult(rs, fmt.Sprintf("%v", rs)).Check(testkit.Rows("1 11"))
+	txn, err = tk.Session().Txn(false)
+	require.NoError(t, err)
+	require.True(t, txn.Valid())
+	tk.MustExec("commit")
+}
+
+func TestTxnGoString(t *testing.T) {
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("use test")
+	tk.MustExec("drop table if exists gostr;")
+	tk.MustExec("create table gostr (id int);")
+
+	txn, err := tk.Session().Txn(false)
+	require.NoError(t, err)
+	require.Equal(t, "Txn{state=invalid}", fmt.Sprintf("%#v", txn))
+
+	tk.MustExec("begin")
+	txn, err = tk.Session().Txn(false)
+	require.NoError(t, err)
+
+	require.Equal(t, fmt.Sprintf("Txn{state=valid, txnStartTS=%d}", txn.StartTS()), fmt.Sprintf("%#v", txn))
+
+	tk.MustExec("insert into gostr values (1)")
+	require.Equal(t, fmt.Sprintf("Txn{state=valid, txnStartTS=%d}", txn.StartTS()), fmt.Sprintf("%#v", txn))
+
+	tk.MustExec("rollback")
+	require.Equal(t, "Txn{state=invalid}", fmt.Sprintf("%#v", txn))
+}
